requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
faiss-cpu
sentence-transformers
python-multipart